Refactor scaling-postgresql module into pipeline-postgresql module
