+++ shardingsphere-scaling/shardingsphere-scaling-dialect/pom.xml
@@ -30,7 +30,6 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>shardingsphere-scaling-postgresql</module>
         <module>shardingsphere-scaling-opengauss</module>
     </modules>
 </project>
